requirements.txt
asgiref==3.7.2
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyMySQL==1.1.0
python-decouple==3.8
python-dotenv==1.0.1
Markdown==3.5.2
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
PyJWT==2.8.0
pillow==10.2.0